Except func.

+++ include/list.h
@@ -9,7 +9,7 @@ struct ListHead {
 typedef struct ListHead ListHead;
 
 #define list_entry(ptr, type, member) \
-	((type*)((char*)(ptr) - (int)(&((type*)0)->member)))
+	((type*)((char*)(ptr) - (long)(&((type*)0)->member)))
 
 static inline void
 list_add(ListHead *prev, ListHead *next, ListHead *data) {
